+++ packages/ts-newskit/src/components/slices/lead-story/styles.ts
@@ -2,8 +2,6 @@ import {
   styled,
   TextBlock,
   Stack,
-  Tag,
-  getColorCssFromTheme,
   Divider,
   getMediaQueryFromTheme
 } from 'newskit';
@@ -16,10 +14,6 @@ export const StyledTextBlock = styled(TextBlock)`
 
 export const StyledTextStack = styled(Stack)`
   border-top: 1px dashed black;
-`;
-
-export const ColouredText = styled(Tag)<{ $color?: string }>`
-  ${({ $color }) => $color && getColorCssFromTheme('color', $color)};
 `;
 
 export const StyledDivider = styled(Divider)`