+++ packages/ts-newskit/src/utils/TCThemeProvider.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
 import { NewsKitProvider, Theme } from 'newskit';
-import { TimesWebLightTheme } from '../theme';
+import { TimesWebLightTheme, TimesWebLightSportTheme } from '../theme';
 
 type TCThemeProviderProps = {
   theme?: Theme;
+};
+
+export const themes = {
+  TimesWebLightTheme,
+  TimesWebLightSportTheme
 };
 
 export const TCThemeProvider: React.FC<TCThemeProviderProps> = ({